requirements.txt
datasets>=2.20
openai>=1.40.0
tqdm>=4.66
python-dotenv>=1.0
google-generativeai>=0.8.0
tenacity>=8.2.3
